chore: save money as long

+++ app/src/main/java/com/finius/core/data/TransactionRepository.kt
@@ -36,7 +36,7 @@ class TransactionRepository(
     fun insert(
         id: String,
         title: String,
-        amount: Double,
+        amount: Long,
         type: TransactionType,
         counterParty: String,
         accountId: String,
